+++ lib/runner.js
@@ -4,6 +4,7 @@ var util = require('util');
 var Promise = require('bluebird');
 var hasFlag = require('has-flag');
 var Test = require('./test');
+var Hook = require('./hook');
 var send = require('./send');
 var objectAssign = require('object-assign');
 
@@ -43,28 +44,6 @@ Runner.prototype.addTest = function (title, cb, opts) {
 	}, opts);
 
 	this.tests.push(test);
-};
-
-function Hook(type, title, fn) {
-	if (!(this instanceof Hook)) {
-		return new Hook(type, title, fn);
-	}
-
-	if (typeof title === 'function') {
-		fn = title;
-		title = null;
-	}
-
-	this.type = type;
-	this.title = title;
-	this.fn = fn;
-}
-
-Hook.prototype.test = function (testTitle) {
-	var title = this.title || (this.type + ' for "' + testTitle + '"');
-	var test = new Test(title, this.fn);
-	test.type = this.type;
-	return test;
 };
 
 Runner.prototype.addSerialTest = function (title, fn) {
